acyclic import (#226)
Была ошибка циклического импорта в plugins

Также исправил опечатку в названии функции
