+++ plugins/github/__init__.py
@@ -1,10 +1,13 @@
-from .fetch import (
+from .fetch_get import (
     fetch_gh_invations,
     fetch_gh_members,
     fetch_gh_org,
     fetch_gh_repos,
     fetch_gh_teams,
+    flatten,
+    get_all_gh_data,
     get_conn,
+    get_gh_data,
+    get_organization,
     upload_df,
 )
-from .get import flatten, get_all_gh_data, get_gh_data, get_organization
